refactor: Move CandidateDao utils to fixtures
